mds_data_sync/mds2mysql/requirements.txt
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
SQLAlchemy==2.0.18
mysql-connector-python==8.0.33
python-dotenv==1.0.0